requirements.txt
python-dotenv 
langchain 
flask 
gunicorn
langchain_openai